Improve performance of the edit_club admin view.
The "members" reverse relation needs a bunch of annotations and
select_related calls to provide the level of detail we want on the page
output, while not blowing out the query count.

This fixes everything in the base templates we ship, *except* for the
permission checking that guardian provides. Still need to add
[Prefetching permissions][p] to resolve this.

[p]: http://django-guardian.readthedocs.io/en/stable/userguide/performance.html#prefetching-permissions