Remove the SportingPulse synchronisation logic.
At best this was a fragile point in time implementation to support the
Touch World Cup in 2015 -- the convening body was using a different
service provider for result entry, and to ensure the International body
had a near-realtime set of results on their own website I implemented a
web scraper to achieve that.

Because we aren't actively maintaining this in light of (potential)
interface changes on the remote service, the value of this is limited.

This change leaves the data source URL on the Division model (for now)
however it's marked editable=False.