Improve Sitemap XML generation in CompetitionSite
Previously we were recursively walking competitions, seasons, divisions, and
matches. In large sites with many of each of the higher order constructs, this
was taking a long time to generate.

This should now be avoided and managed through a single query on the lowest
level construct (the Match) with select_related traversing all the
relationships back up to the highest (Competition).