Allow TOUCHTECHNOLOGY_PAGE_CONTENT_BLOCKS to be path to callback.
Makes it easy to provide dynamic behaviour for the
TOUCHTECHNOLOGY_PAGE_CONTENT_BLOCKS setting. Passing in the `node`
allows us to have a variable number of blocks for some pages (example
might be for a home page over "normal" pages, or a multilingual site
that wants to provide different content for different audiences).