Add notion of opponent_division to LadderEntry to be used in RankPoints creation.
When we have matches that are played across ranking divisions they
should not be included in either team's ranking. For FIT this is a
diminishing factor (basically we're moving to eradicate combined
divisions) however there are some historical cases in scope.

Also added a --rebuild option to the produce_rank management command so
that we can easily regenerate based on existing dates in the database.