Fix failure to call the YouTube API in `edit_match` admin view
In #97 I naively added a check that would stop the `pre_save_callback` being executed when the value of `Match.live_stream` was false.

However, this is the pre-save value. At the time we evaluate it, the form has not been processed. Thus, a match that is transitioning from `False` to `True` will never fire the callback, so the YouTube resource is never created.

This is still sub-optimal and may cause other issues (see #85), but I don't have time to spend on it now.